requirements.txt
trino
requests-oauthlib
python-dotenv
pandas
pyarrow
tqdm